Added Karma config (for mocha tests, and coverage)
After **a lot** of testing multiple solutions, this
`istanbul-instrumenter-loader` based version has almost all my
requirements. The main thing missing is that includeAllSources doesn't
actually include all sources, so modules that don't have a .tests.js
file aren't shown as not covered :( This was also an issue back with
`nyc` when I had to transpile, so it's not really a new issue, but at
least with nyc it was fixed eventually, so it sucks that this issue is
back.